+++ dados/gold/pa_indexadores_producao_rural/censo_2006_2017_despesas.py
@@ -34,7 +34,7 @@ def extract() -> pd.DataFrame:
             tipo_agricultura,
             tipo_despesa,
             quantidade_estabelecimentos_fizeram_despesa,
-            ROUND(valor_despesa, 2) AS valor_despesa
+            ROUND(valor_despesa::numeric, 2) AS valor_despesa
         FROM {SILVER_SCHEMA}.{SILVER_TABLE_2006}
         WHERE id_municipio LIKE '15%'
         UNION ALL
@@ -44,7 +44,7 @@ def extract() -> pd.DataFrame:
             tipo_agricultura,
             tipo_despesa,
             quantidade_estabelecimentos_fizeram_despesa,
-            ROUND(valor_despesa, 2) AS valor_despesa
+            ROUND(valor_despesa::numeric, 2) AS valor_despesa
         FROM {SILVER_SCHEMA}.{SILVER_TABLE_2017}
         WHERE id_municipio LIKE '15%'
     """
