+++ dados/gold/pa_indexadores_producao_rural/censo_2017_lavoura_permanente.py
@@ -40,11 +40,11 @@ def extract() -> pd.DataFrame:
             quantidade_produzida,
             quantidade_vendida,
             quantidade_vendida AS comercio_quantidade_produzida,
-            quantidade_produzida - quantidade_vendida AS autoconsumo_quantidade_vendida,
+            quantidade_produzida::numeric - quantidade_vendida::numeric AS autoconsumo_quantidade_vendida,
             valor_producao,
             valor_venda,
             valor_venda AS comercio_valor_producao,
-            valor_producao - valor_venda AS autoconsumo_valor_producao,
+            valor_producao::numeric - valor_venda::numeric AS autoconsumo_valor_producao,
             area_colhida,
             area_plantada
         FROM {SILVER_SCHEMA}.{SILVER_TABLE}
